In hindsight, ProtocolVersion should be a UInt16 because of rapid development (ironically this change breaks protocol, exactly what versioning is supposed to prevent, duh). Also teams will technically be a player too in messaging (it's how BZFlag does it), so teams should start at 240 so that slot IDs for players can occur below that.

+++ AngryTanks.Client/ServerLink.cs
@@ -38,7 +38,7 @@ namespace AngryTanks.Client
 
             // TODO be able to change callsign/tag
             hailMessage.Write((byte)MessageType.MsgEnter);
-            hailMessage.Write((byte)Protocol.ProtocolVersion);
+            hailMessage.Write(Protocol.ProtocolVersion);
             hailMessage.Write((byte)TeamType.RogueTeam);
             hailMessage.Write("Player Callsign");
             hailMessage.Write("Player Tag");
